+++ src/datasources/ascii/asciidatareader.h
@@ -31,16 +31,16 @@ class AsciiDataReader
     ~AsciiDataReader();
 
     void clear();
-    void setRow0Begin(int begin);
-    inline int beginOfRow(int row) const { return _rowIndex[row]; }
-    inline int numberOfFrames() const { return _numFrames; }
+    void setRow0Begin(qint64 begin);
+    inline qint64 beginOfRow(qint64 row) const { return _rowIndex[row]; }
+    inline qint64 numberOfFrames() const { return _numFrames; }
 
     // where
     const AsciiFileBuffer::RowIndex& rowIndex() const { return _rowIndex; }
     
     void detectLineEndingType(QFile& file);
     
-    bool findDataRows(bool read_completely, QFile& file, int _byteLength);
+    bool findDataRows(bool read_completely, QFile& file, qint64 _byteLength);
     int readField(const AsciiFileData &buf, int col, double *v, const QString& field, int s, int n);
     int readFieldFromChunk(const AsciiFileData& chunk, int col, double *v, const QString& field);
 
@@ -57,24 +57,24 @@ class AsciiDataReader
     const AsciiCharacterTraits::IsWhiteSpace isWhiteSpace;
 
     template<class T>
-    bool resizeBuffer(T& buffer, int bytes);
+    bool resizeBuffer(T& buffer, qint64 bytes);
 
     template<class Buffer, typename ColumnDelimiter>
-    int readColumns(double* v, const Buffer& buffer, int bufstart, int bufread, int col, int s, int n,
+    int readColumns(double* v, const Buffer& buffer, qint64 bufstart, qint64 bufread, int col, int s, int n,
                     const AsciiCharacterTraits::LineEndingType&, const ColumnDelimiter&) const;
 
     template<class Buffer, typename ColumnDelimiter, typename CommentDelimiter>
-    int readColumns(double* v, const Buffer& buffer, int bufstart, int bufread, int col, int s, int n,
+    int readColumns(double* v, const Buffer& buffer, qint64 bufstart, qint64 bufread, int col, int s, int n,
                     const AsciiCharacterTraits::LineEndingType&, const ColumnDelimiter&, const CommentDelimiter&) const;
 
     template<class Buffer, typename IsLineBreak, typename ColumnDelimiter, typename CommentDelimiter, typename ColumnWidthsAreConst>
-    int readColumns(double* v, const Buffer& buffer, int bufstart, int bufread, int col, int s, int n,
+    int readColumns(double* v, const Buffer& buffer, qint64 bufstart, qint64 bufread, int col, int s, int n,
                     const IsLineBreak&, const ColumnDelimiter&, const CommentDelimiter&, const ColumnWidthsAreConst&) const;
 
     template<class Buffer, typename IsLineBreak, typename CommentDelimiter>
-    bool findDataRows(const Buffer& buffer, int bufstart, int bufread, const IsLineBreak&, const CommentDelimiter&);
+    bool findDataRows(const Buffer& buffer, qint64 bufstart, qint64 bufread, const IsLineBreak&, const CommentDelimiter&);
 
-    void toDouble(const LexicalCast& lexc, const char* buffer, int bufread, int ch, double* v, int row) const;
+    void toDouble(const LexicalCast& lexc, const char* buffer, qint64 bufread, qint64 ch, double* v, int row) const;
 
     mutable QMutex _localeMutex;
 };
